refactor: UserService의 session로직 컨트롤러로 이동
